Add pseudo-yaml parsing of RFM69 packets using jsmn
The zserge/jsmn json parser, when running in non-strict mode (the
default) is happy to parse something that looks like yaml:

{f: {water_meter: {litres: 1234, tamper: 0}}}

The unquoted tokens are classed as 'primitives' but are accepted.

Add a 'yparse' wrapper around this library to provide parsing
of incoming packets, provide a json emitter, and coerce
unrecognised primitives to strings.

does not insist that strings are quoted (it parses them
as primitives